requirements.txt
# Python
psycopg2==2.8.6
requests==2.25.1
ipython==7.16.1

celery==4.4.1
redis==3.5.3
Pillow==8.1.0

python-decouple==3.4
python-dateutil==2.8.1


# Django
Django==3.2
django-cors-headers==3.7.0
django-model-utils==4.1.1

django-extensions==3.1.0
django-filter==2.4.0

# Django Rest
djangorestframework==3.12.2

# Deploy
gunicorn==20.0.4

# Swagger
coreapi==2.3.3
openapi-codec==1.3.2

